Add files via upload
I had to take out the "newline" to make it work for me.  Also, I'm sure that my queries are very inelegant but I hope I did it somewhat correctly.  Also we both need to create the grocery.db database again because I forgot to include bread in the NOT IN when I was finding the non probability items.  The bread count was higher than that of milk because of my mistake.